Move dithering to common module
