+++ zeapp/android/src/main/java/de/berlindroid/zeapp/zeui/ZeCameraEditor.kt
@@ -20,7 +20,7 @@ import com.commit451.coiltransformations.CropTransformation
 import de.berlindroid.zeapp.BuildConfig
 import de.berlindroid.zeapp.PAGE_HEIGHT
 import de.berlindroid.zeapp.PAGE_WIDTH
-import de.berlindroid.zeapp.zebits.ditherFloydSteinberg
+import de.berlindroid.zekompanion.ditherFloydSteinberg
 import de.berlindroid.zeapp.zemodels.ZeConfiguration
 import de.berlindroid.zeapp.zemodels.ZeEditor
 import de.berlindroid.zeapp.zevm.ZeBadgeViewModel
